Workaround Tera's group_by limitation
The group_by was converting the month integers into strings, which meant
October (10) was being ordered before December (12) after sorting.

To resolve, I stopped grouping on the month and instead sorted by the
month integer value. This then required an update to the conditional
logic to support providing state for determining when to put a new Month
+ Year header as well as ending/starting unordered lists/divs for
  styling purposes.

This means a bit more state and more convoluted conditionals but it
works around the group_by issues for now in pure template form. An
alternative could have been providing a totally different structure
within the template context, with posts already grouped.