fix(archives): avoid group_by entirely
Like the previous fix (574d3cd002) where group_by was turning months
into a string, resulting in string comparisons on months instead of a
integer comparison during the sort, I was hitting the same problem with
the year value as well. I noticed this when importing some posts from
2010 into the new service.

This change fixes the archive behavior to show posts in reverse
chronological order by assuming the posts provided are already sorted.
We then reverse them and add a marker current_year variable for figuring
out when to reset our already existing current_month marker. These two
variables are used for determining when to add a new Year, Month header
and end the unordered list of posts from that month.

Additionally, the previous scheme was resulting in posts not being
sorted by day of month within each block. This is also now resolved.

Refs: 574d3cd002