feat: add bookmark-of support
For archives, display a bookmark emoji (🔖) before the title but doesn't
modify the link to the post in the article page. On article pages, make
the title link to the bookmark rather than be a permalink to the post.
Makes the "posted at" permalink at the bottom always annotated as a
permalink.